Changed types and removed istream.peek()

+++ erohin.vladimir/T1/complexquad.cpp
@@ -9,7 +9,7 @@ erohin::Complexquad::Complexquad(point_t corner1, point_t corner2, point_t corne
   vertex_{ corner1, corner2, corner3, corner4 }
 {
   point_t center = findIntersectionPoint(vertex_);
-  for (int i = 0; i < 3; i += 2)
+  for (size_t i = 0; i < 3; i += 2)
   {
     if (!isPointOnSegment(center, vertex_[i], vertex_[i + 1]))
     {
@@ -25,7 +25,7 @@ double erohin::Complexquad::getArea() const
   double area = 0.0;
   point_t point[3]{ 0.0, 0.0 };
   point[0] = getCenter();
-  for (int i = 0; i < 2; ++i)
+  for (size_t i = 0; i < 2; ++i)
   {
     point[1] = vertex_[i];
     point[2] = vertex_[3 - i];
@@ -46,7 +46,7 @@ erohin::point_t erohin::Complexquad::getCenter() const
 
 void erohin::Complexquad::move(double dx, double dy)
 {
-  for (int i = 0; i < 4; ++i)
+  for (size_t i = 0; i < 4; ++i)
   {
     vertex_[i].x += dx;
     vertex_[i].y += dy;
@@ -68,7 +68,7 @@ void erohin::Complexquad::scale(double ratio)
     throw std::invalid_argument("Wrong figure creation");
   }
   point_t center = getCenter();
-  for (int i = 0; i < 4; ++i)
+  for (size_t i = 0; i < 4; ++i)
   {
     vertex_[i].x = center.x + (vertex_[i].x - center.x) * ratio;
     vertex_[i].y = center.y + (vertex_[i].y - center.y) * ratio;
